+++ tests/testthat/test_recency_stats.R
@@ -3,11 +3,13 @@ library(remstats)
 
 test_that("recencyContinue", {
 	
+	# Compute the statistics
 	effects <- ~ recencyContinue()
 	tomres <- tomstats(effects, edgelist = history)
-	aomres <- aomstats(edgelist = history, receiver_effects = effects)
+	aomres <- aomstats(edgelist = history[,c(1:3)], receiver_effects = effects)
 	
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
+	# Test whether the statistic for this event is correct 
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
 		past <- which(tomres$evls[past,1] == event)
@@ -17,19 +19,20 @@ test_that("recencyContinue", {
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 	
-	sender <- tomres$riskset$actor1[9]
-	receivers <- sort(unique(tomres$riskset$actor2))
-	receiver <- match(tomres$riskset$actor2[9], receivers)
+	sender <- tomres$riskset[event, 1]
+	receivers <- sort(unique(tomres$riskset[,2]))
+	receiver <- match(tomres$riskset[event, 2], receivers)
 	events <- which(history$actor1 == sender)
-	expect_true(all(aomres$statistics$receiver_stats[events,10,] %in% 
-			tomres$statistics[,9,2]))
+	expect_true(all(aomres$statistics$receiver_stats[events, receiver,] %in% 
+			tomres$statistics[,event,2]))
 	
+	# Test for consider_type
 	colnames(history)[4] <- "type"
 	effects <- ~ recencyContinue(consider_type = TRUE)
 	tomres <- tomstats(effects, edgelist = history)
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
 		past <- which(tomres$evls[past,1] == event)
@@ -39,165 +42,165 @@ test_that("recencyContinue", {
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 })
 
 test_that("recencySendSender", {
 	
 	effects <- ~ recencySendSender()
 	tomres <- tomstats(effects, edgelist = history)
-	aomres <- aomstats(edgelist = history, sender_effects = effects)
+	aomres <- aomstats(edgelist = history[,c(1:3)], sender_effects = effects)
 	
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,2] == tomres$riskset$actor1[9])
+		past <- which(tomres$edgelist[past,2] == tomres$riskset[event,1])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 	
-	senders <- sort(unique(tomres$riskset$actor1))
-	sender <- match(tomres$riskset$actor1[9], senders)
+	senders <- sort(unique(tomres$riskset[,1]))
+	sender <- match(tomres$riskset[event, 1], senders)
 	expect_true(all(aomres$statistics$sender_stats[,sender,2] %in% 
-			tomres$statistics[,9,2]))
+			tomres$statistics[,event,2]))
 	
 	colnames(history)[4] <- "type"
 	effects <- ~ recencySendSender(consider_type = TRUE)
 	tomres <- tomstats(effects, edgelist = history)
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,2] == tomres$riskset$actor1[9] & 
-				tomres$edgelist[past,4] == tomres$riskset$type[9])
+		past <- which(tomres$edgelist[past,2] == tomres$riskset[event,1] & 
+				tomres$edgelist[past,4] == tomres$riskset[event,3])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 })
 
 test_that("recencySendReceiver", {
 	
 	effects <- ~ recencySendReceiver()
 	tomres <- tomstats(effects, edgelist = history)
-	aomres <- aomstats(edgelist = history, receiver_effects = effects)
+	aomres <- aomstats(edgelist = history[,c(1:3)], receiver_effects = effects)
 	
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,2] == tomres$riskset$actor2[9])
+		past <- which(tomres$edgelist[past,2] == tomres$riskset[event,2])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 	
-	receivers <- sort(unique(tomres$riskset$actor2))
-	receiver <- match(tomres$riskset$actor2[9], receivers)
+	receivers <- sort(unique(tomres$riskset[,2]))
+	receiver <- match(tomres$riskset[event,2], receivers)
 	expect_true(all(aomres$statistics$receiver_stats[,receiver,1] %in% 
-			tomres$statistics[,9,2]))
+			tomres$statistics[,event,2]))
 	
 	colnames(history)[4] <- "type"
 	effects <- ~ recencySendReceiver(consider_type = TRUE)
 	tomres <- tomstats(effects, edgelist = history)
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,2] == tomres$riskset$actor2[9] & 
-				tomres$edgelist[past,4] == tomres$riskset$type[9])
+		past <- which(tomres$edgelist[past,2] == tomres$riskset[event,2] & 
+				tomres$edgelist[past,4] == tomres$riskset[event,3])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 })
 
 test_that("recencyReceiveSender", {
 	
 	effects <- ~ recencyReceiveSender()
 	tomres <- tomstats(effects, edgelist = history)
-	aomres <- aomstats(edgelist = history, sender_effects = effects)
+	aomres <- aomstats(edgelist = history[,c(1:3)], sender_effects = effects)
 	
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,3] == tomres$riskset$actor1[9])
+		past <- which(tomres$edgelist[past,3] == tomres$riskset[event,1])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 	
-	senders <- sort(unique(tomres$riskset$actor1))
-	sender <- match(tomres$riskset$actor1[9], senders)
+	senders <- sort(unique(tomres$riskset[,1]))
+	sender <- match(tomres$riskset[event,1], senders)
 	expect_true(all(aomres$statistics$sender_stats[,sender,2] %in% 
-			tomres$statistics[,9,2]))
+			tomres$statistics[,event,2]))
 	
 	colnames(history)[4] <- "type"
 	effects <- ~ recencyReceiveSender(consider_type = TRUE)
 	tomres <- tomstats(effects, edgelist = history)
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,3] == tomres$riskset$actor1[9] & 
-				tomres$edgelist[past,4] == tomres$riskset$type[9])
+		past <- which(tomres$edgelist[past,3] == tomres$riskset[event,1] & 
+				tomres$edgelist[past,4] == tomres$riskset[event,3])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 })
 
 test_that("recencyReceiveReceiver", {
 	
 	effects <- ~ recencyReceiveReceiver()
 	tomres <- tomstats(effects, edgelist = history)
-	aomres <- aomstats(edgelist = history, receiver_effects = effects)
+	aomres <- aomstats(edgelist = history[,c(1:3)], receiver_effects = effects)
 	
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,3] == tomres$riskset$actor2[9])
+		past <- which(tomres$edgelist[past,3] == tomres$riskset[event, 2])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 	
-	receivers <- sort(unique(tomres$riskset$actor2))
-	receiver <- match(tomres$riskset$actor2[9], receivers)
+	receivers <- sort(unique(tomres$riskset[,2]))
+	receiver <- match(tomres$riskset[event, 2], receivers)
 	expect_true(all(aomres$statistics$receiver_stats[,receiver,1] %in% 
-			tomres$statistics[,9,2]))
+			tomres$statistics[,event,2]))
 	
 	colnames(history)[4] <- "type"
 	effects <- ~ recencyReceiveReceiver(consider_type = TRUE)
 	tomres <- tomstats(effects, edgelist = history)
-	event <- 9
+	event <- sample(1:nrow(tomres$riskset), 1) # randomly sample an event
 	expect_equal(sapply(1:nrow(history), function(m) {
 		past <- which(history$time < history$time[m])
-		past <- which(tomres$edgelist[past,3] == tomres$riskset$actor2[9] & 
-				tomres$edgelist[past,4] == tomres$riskset$type[9])
+		past <- which(tomres$edgelist[past,3] == tomres$riskset[event, 2] & 
+				tomres$edgelist[past,4] == tomres$riskset[event ,3])
 		if(length(past)>0) {
 			last <- max(past)
 			1/((history$time[m] - history$time[last])+1)	
 		} else {
 			0
 		}
-	}), tomres$statistics[,9,2])
+	}), tomres$statistics[,event,2])
 })