+++ NEWS.md
@@ -1,3 +1,15 @@
+# remstats 3.0.0
+Date: 22-Dec-21
+
+IMPORTANT: remstats 3.0.0 works with remify >= 2.0.0. Make sure remify is updated. If earlier versions of remify are in use, remstats 3.0.0 will break down. 
+
+## Major changes
+* Integrated new version (2.0.0) of remify. 
+
+## Minor changes
+* Fixed bug in `tie()` (see issue [#29](https://github.com/TilburgNetworkGroup/remstats/issues/29)).
+* Added warning messages when redundant actors are present in input data for the exogenous stats (see issue [#30](https://github.com/TilburgNetworkGroup/remstats/issues/30)).
+
 # remstats 2.0.4
 Date: 21-Dec-21